feat: Implement game actions handling
This commit implements the handling of game actions on the PlayTable screen.

Key changes include:

- Added GameActionsBatch
DTO and GameActionDTO to represent game actions received from Firebase.
- Added GameActionsBatch and GameAction classes to represent game actions in the domain layer.
- Implemented getGameActionsFlow in GameSessionRepository to fetch game actions from Firebase.
- Added logic to PlayTableViewModel to handle game actions and update
 the PlayTableState accordingly.
- Added shimmer loading animation to CreateNewSessionScreen while waiting for game session data.
- Updated PlayTableScreen to use Long for card IDs instead of Int.
- Updated PlayTableViewModel to use Long for card IDs instead of Int.
- Updated DataHolder to be a common model.

+++ app/src/main/java/com/hanas/addy/repository/gemini/SampleResponse.kt
@@ -230,6 +230,6 @@ val samplePlayCardStack = samplePlayCardStackJsonResponse
     .mapToPlayCardStack("")
     .let {
         it.copy(cards = it.cards.mapIndexed { id, data ->
-            data.copy(id = id)
+            data.copy(id = id.toLong())
         })
     }
